Pass context to text nodes
